Changing the time units to default (nanoseconds) in timeLib Testing the library using exercise 2.2.3

+++ 00_Lib/timeLib.h
@@ -13,7 +13,7 @@ class timeLib
         timeLib();
         void startMeasurement(void);
         void stopMeasurement(void);
-        int getMeasurement(void);
+        long getMeasurement(void);
 };
 
 #endif // TIME_LIB_H
